Fixed off-by-one error in the segment inequality constraints.
Previously, the delta-time was erroneously determined by subtracting two from
the number of intermediate segment points. The result was that the total time
would exceed 1.0, violating the time constraint. Changed from -2 to +2 - 1 to
account for the two endpoints and the necessary conversion from number of
waypoints to number of intermediate segments. Added a comment to clarify this in
the code.